+++ molo/core/templatetags/core_tags.py
@@ -7,7 +7,8 @@ from markdown import markdown
 
 from molo.core.models import (
     Page, ArticlePage, SectionPage, SiteSettings, Languages, Tag,
-    ArticlePageTags, SectionIndexPage)
+    ArticlePageTags, SectionIndexPage, ReactionQuestion,
+    ReactionQuestionChoice)
 
 register = template.Library()
 
@@ -331,6 +332,49 @@ def get_articles_for_tags_with_translations(
 
 
 @register.assignment_tag(takes_context=True)
+def get_articles_for_tag(context, tag):
+    request = context['request']
+    locale = context.get('locale_code')
+    if tag:
+        main_tag = tag.specific.get_main_language_page()
+        pks = [article_tag.page.pk for article_tag in
+               ArticlePageTags.objects.filter(tag=main_tag)]
+        return get_pages(
+            context, ArticlePage.objects.descendant_of(
+                request.site.root_page).filter(pk__in=pks), locale)
+    return None
+
+
+@register.assignment_tag(takes_context=True)
+def get_next_tag(context, tag):
+    request = context['request']
+    locale_code = context.get('locale_code')
+    if tag:
+        main_tag = tag.get_main_language_page()
+        if main_tag:
+            if main_tag.get_next_sibling() and \
+                    main_tag.get_next_sibling().languages.filter(
+                        language__is_main_language=True).exists():
+                next_tag = main_tag.get_next_sibling()
+                if next_tag.specific.get_translation_for(
+                        locale_code, context['request'].site):
+                    return next_tag.specific.get_translation_for(
+                        locale_code, context['request'].site)
+                else:
+                    return next_tag
+            else:
+                next_tag = Tag.objects.descendant_of(
+                    request.site.root_page).filter(
+                    languages__language__is_main_language=True).live().first()
+                if next_tag.specific.get_translation_for(
+                        locale_code, context['request'].site):
+                    return next_tag.specific.get_translation_for(
+                        locale_code, context['request'].site)
+                else:
+                    return next_tag
+
+
+@register.assignment_tag(takes_context=True)
 def get_tags_for_section(context, section, tag_count=2, tag_article_count=4):
     request = context['request']
     locale = context.get('locale_code')
@@ -455,6 +499,43 @@ def load_tags_for_article(context, article):
     else:
         return []
     return get_pages(context, qs, locale)
+
+
+@register.assignment_tag(takes_context=True)
+def load_choices_for_reaction_question(context, question):
+    locale = context.get('locale_code')
+    if question:
+        question_pk = question.get_main_language_page().pk
+        question = ReactionQuestion.objects.filter(pk=question_pk)
+    if question and question.first().get_children():
+        pks = [c.pk for c in question.first().get_children().filter(
+            languages__language__is_main_language=True)]
+        choices = ReactionQuestionChoice.objects.filter(pk__in=pks)
+        return get_pages(context, choices, locale)
+    return []
+
+
+@register.assignment_tag(takes_context=True)
+def load_reaction_question(context, article):
+    locale = context.get('locale_code')
+    request = context['request']
+    question = None
+    if article:
+        article_question = article.get_main_language_page() \
+            .specific.reaction_questions.all().first()
+        if hasattr(article_question, 'reaction_question'):
+            question = article_question.reaction_question
+
+        if question and request.site:
+            qs = ReactionQuestion.objects.descendant_of(
+                request.site.root_page).live().filter(
+                    pk=question.pk, languages__language__is_main_language=True)
+        else:
+            return []
+        translated_question = get_pages(context, qs, locale)
+        if translated_question:
+            return get_pages(context, qs, locale)[0]
+        return question
 
 
 @register.assignment_tag(takes_context=True)